requirements.txt
requests>=2.31.0
aiohttp>=3.8.0
pyyaml>=6.0 
openai>=1.0.0
pydantic>=1.10.0
# 如需支持 MongoDB/PostgreSQL 可后续补充 